Refatorando rotas, Party model, parties, party Controller

+++ backend/controllers/partyController.js
@@ -1,11 +1,14 @@
 import { response } from "express";
-import Party from "../models/Party.js";
+import { Party } from "../models/Party.js";
 import mongoose from "mongoose";
+import User from "../models/User.js";
+import getToken from "../helpers/get-token.js";
+import getUserByToken from "../helpers/get-user-by-token.js";
 
-const checkPartyBudget = (budget, services) =>{
+const checkPartyBudget = (budget, services) => {
     const priceSum = services.reduce((sum, service) => sum + service.price, 0);
 
-    if(priceSum > budget) return false;
+    if (priceSum > budget) return false;
 
     return true;
 
@@ -13,41 +16,76 @@ const checkPartyBudget = (budget, services) =>{
 
 const partyController = {
     // Função para criar party
-    create: async (req, res) =>{
+    create: async (req, res) => {
         try {
+
+            const token = getToken(req);
+
+            if (!token) {
+                res.status(401).json({ msg: "Token inválido" });
+                return;
+            }
+
+            const user = await getUserByToken(token);
+
+            if (!user) {
+                res.status(401).json({ msg: "Token inválido" });
+                return;
+            }
+
             const party = {
                 title: req.body.title,
                 author: req.body.author,
                 description: req.body.description,
                 budget: req.body.budget,
                 image: req.body.image,
-                services: req.body.services
+                services: req.body.services,
+                user: {
+                    _id: user._id,
+                    name: user.name,
+                    image: user.image,
+                    phone: user.phone,
+                },
             }
 
-            if(party.services && !checkPartyBudget(party.budget, party.services)){
-                res.status(406).json({msg: "O seu orçamento é insuficiente"});
+            if (party.services && !checkPartyBudget(party.budget, party.services)) {
+                res.status(406).json({ msg: "O seu orçamento é insuficiente" });
                 return;
             }
 
             const response = await Party.create(party);
 
-            res.status(201).json({response ,msg: 'Festa criada com sucesso' });
+            res.status(201).json({ response, msg: 'Festa criada com sucesso' });
 
         } catch (error) {
             console.log(error);
         }
     },
-    // Função para resgatar todas as festas
-    getAll: async (req, res) =>{
+    // Função para resgatar todas as festas do usuário
+    getAll: async (req, res) => {
         try {
-            const parties = await Party.find();
-            res.status(200).json({parties ,msg:"Dados de festas recuperados com sucesso!"})
+            const token = getToken(req)
+
+            if (!token) {
+                res.status(401).json({ msg: "Token inválido" });
+                return;
+            }
+
+            const user = await getUserByToken(token)
+
+            if (!user) {
+                res.status(401).json({ msg: "Token inválido" });
+                return;
+            }
+
+            const parties = await Party.find({ 'user._id': user._id });
+            res.status(200).json({ parties, msg: "Dados de festas recuperados com sucesso!" })
         } catch (error) {
             console.log(error);
         }
     },
 
-    // Função para resgatar uma festa pelo ID
+    // Função para resgatar uma festa pelo ID do usuário
     get: async (req, res) => {
         try {
             const id = req.params.id;
@@ -55,11 +93,32 @@ const partyController = {
             if (!mongoose.Types.ObjectId.isValid(id)) {
                 return res.status(400).json({ msg: 'ID inválido' });
             }
+
+            const token = getToken(req)
+
+            if (!token) {
+                res.status(401).json({ msg: "Token inválido" });
+                return;
+            }
+            const user = await getUserByToken(token)
+            if (!user) {
+                res.status(401).json({ msg: "Token inválido" });
+                return;
+            }
+
             const party = await Party.findById(id);
 
             if (!party) {
                 res.status(404).json({ msg: "Festa não encontrada." });
                 return;
+            }
+
+            if (party.user._id.toString() != user._id.toString()) {
+                res.status(404).json({
+                    message:
+                        'Houve um problema em processar sua solicitação, tente novamente mais tarde!',
+                })
+                return
             }
 
             res.status(200).json({ party, msg: "Festa recuperada com sucesso" });
@@ -71,7 +130,7 @@ const partyController = {
     delete: async (req, res) => {
         try {
             const id = req.params.id;
-            
+
             if (!mongoose.Types.ObjectId.isValid(id)) {
                 return res.status(400).json({ msg: 'ID inválido' });
             }
@@ -82,6 +141,28 @@ const partyController = {
                 res.status(404).json({ msg: "Festa não encontrada." });
                 return;
             }
+
+            const token = getToken(req)
+
+            if (!token) {
+                res.status(401).json({ msg: "Token inválido" });
+                return;
+            }
+            const user = await getUserByToken(token)
+
+            if (!user) {
+                res.status(401).json({ msg: "Token inválido" });
+                return;
+            }
+
+            if (party.user._id.toString() != user._id.toString()) {
+                res.status(404).json({
+                    message:
+                        'Houve um problema em processar sua solicitação, tente novamente mais tarde!',
+                })
+                return
+            }
+
             const deletedParty = await Party.findByIdAndDelete(id);
 
             res.status(200).json({ deletedParty, msg: "Festa deletada com sucesso." });
@@ -94,9 +175,22 @@ const partyController = {
     update: async (req, res) => {
         try {
             const id = req.params.id;
-            
+
             if (!mongoose.Types.ObjectId.isValid(id)) {
                 return res.status(400).json({ msg: 'ID inválido' });
+            }
+
+            const token = getToken(req)
+
+            if (!token) {
+                res.status(401).json({ msg: "Token inválido" });
+                return;
+            }
+            const user = await getUserByToken(token)
+
+            if (!user) {
+                res.status(401).json({ msg: "Token inválido" });
+                return;
             }
 
             const party = {
@@ -105,11 +199,17 @@ const partyController = {
                 description: req.body.description,
                 budget: req.body.budget,
                 image: req.body.image,
-                services: req.body.services
+                services: req.body.services,
+                user: {
+                    _id: user._id,
+                    name: user.name,
+                    image: user.image,
+                    phone: user.phone,
+                },
             }
 
-            if(party.services &&!checkPartyBudget(party.budget, party.services)){
-                res.status(406).json({msg: "O seu orçamento é insuficiente"});
+            if (party.services && !checkPartyBudget(party.budget, party.services)) {
+                res.status(406).json({ msg: "O seu orçamento é insuficiente" });
                 return;
             }
 
@@ -120,7 +220,7 @@ const partyController = {
                 return;
             }
 
-            res.status(200).json({party, msg: "Dados da festa atualizados com sucesso!"});
+            res.status(200).json({ party, msg: "Dados da festa atualizados com sucesso!" });
 
         } catch (error) {
             console.log(error);
